feat(projects): adds the posibility to add multiple icons to a project, each with their own link
Previously, only a single icon could be used for a project. Now multiple icons can be displayed
alongside one icon. In addition, now links can be added to these icons as well.

BREAKING CHANGE: Removes "icon" field on projects and instead uses "icons". To migrate, rename icon
and make the string value the only key for an object. The value of this key should be the empty
string ("").